update recentlyViewedNamespaces on change from list
